+++ src/Microsoft.SqlTools.ServiceLayer/Connection/Contracts/ConnectParamsExtensions.cs
@@ -27,23 +27,4 @@ namespace Microsoft.SqlTools.ServiceLayer.Connection.Contracts
             );
         }
     }
-
-    /// <summary>
-    /// Extension methods to ConnectionSummary
-    /// </summary>
-    public static class ConnectionSummaryExtensions
-    {
-        /// <summary>
-        /// Create a copy of a ConnectionSummary object
-        /// </summary>
-        public static ConnectionSummary Clone(this ConnectionSummary summary)
-        {
-            return new ConnectionSummary()
-            {
-                ServerName = summary.ServerName,
-                DatabaseName = summary.DatabaseName,
-                UserName = summary.UserName
-            };
-        }
-    }
 }
