+++ src/Microsoft.SqlTools.ServiceLayer/Connection/Contracts/ConnectionChangedParams.cs
@@ -3,8 +3,6 @@
 // Licensed under the MIT license. See LICENSE file in the project root for full license information.
 //
 
-using Microsoft.SqlTools.ServiceLayer.Hosting.Protocol.Contracts;
-
 namespace Microsoft.SqlTools.ServiceLayer.Connection.Contracts
 {
     /// <summary>
@@ -22,15 +20,4 @@ namespace Microsoft.SqlTools.ServiceLayer.Connection.Contracts
         /// </summary>
         public ConnectionSummary Connection { get; set; }
     }
-
-    /// <summary>
-    /// ConnectionChanged notification mapping entry 
-    /// </summary>
-    public class ConnectionChangedNotification
-    {
-        public static readonly
-            EventType<ConnectionChangedParams> Type =
-            EventType<ConnectionChangedParams>.Create("connection/connectionchanged");
-    }
-
 }
