+++ src/Microsoft.SqlTools.ServiceLayer/Connection/Contracts/DisconnectParams.cs
@@ -3,8 +3,6 @@
 // Licensed under the MIT license. See LICENSE file in the project root for full license information.
 //
 
-using Microsoft.SqlTools.ServiceLayer.Hosting.Protocol.Contracts;
-
 namespace Microsoft.SqlTools.ServiceLayer.Connection.Contracts
 {
     /// <summary>
@@ -17,15 +15,5 @@ namespace Microsoft.SqlTools.ServiceLayer.Connection.Contracts
         /// or a virtual file representing an object in a database.         
         /// </summary>
         public string OwnerUri { get; set; }
-    }
-
-    /// <summary>
-    /// Disconnect request mapping entry 
-    /// </summary>
-    public class DisconnectRequest
-    {
-        public static readonly
-            RequestType<DisconnectParams, bool> Type =
-            RequestType<DisconnectParams, bool>.Create("connection/disconnect");
     }
 }
